OTel upgrade to 1.31.0
